Most of code port from mock module to ICS 20
