--- x/ibc/mock/bank/internal/types/expected_keepers.go
@@ -1,25 +1,0 @@
-package types
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/ibc/04-channel/exported"
-	transfer "github.com/cosmos/cosmos-sdk/x/ibc/20-transfer"
-	commitment "github.com/cosmos/cosmos-sdk/x/ibc/23-commitment"
-)
-
-// IbcBankKeeper expected IBC transfer keeper
-type IbcBankKeeper interface {
-	ReceiveTransfer(ctx sdk.Context, srcPort, srcChannel, destPort, destChannel string, data transfer.PacketData) error
-}
-
-// ChannelKeeper expected IBC channel keeper
-type ChannelKeeper interface {
-	RecvPacket(
-		ctx sdk.Context,
-		packet exported.PacketI,
-		proof commitment.ProofI,
-		proofHeight uint64,
-		acknowledgement []byte,
-		portCapability sdk.CapabilityKey,
-	) (exported.PacketI, error)
-}
