--- x/ibc/mock/bank/internal/types/keys.go
@@ -1,15 +1,0 @@
-package types
-
-const (
-	// ModuleName is the name of the module
-	ModuleName = "ibcmockbank"
-
-	// StoreKey is the string store representation
-	StoreKey = ModuleName
-
-	// QuerierRoute is the querier route for the module
-	QuerierRoute = ModuleName
-
-	// RouterKey is the msg router key for the module
-	RouterKey = ModuleName
-)
